refactor MainMenu into separate handler classes
